[Sema] Requires nointerpolation on GetAttributeAtVertex param #0
This commit adds new sema checks to restrict the scope of the feature:
the `nointerpolation` attribute is now required on all parameters used
in GetAttributeAtVertex, and doesn't propagage magically to parent/child
functions.

Those issues are mostly caused by the feature which is badly designed
at the language/spec level. The SPIR-V side crashes during validation,
which is a bit unfortunate as the error message is not trivial to
understand. For this reason, I moved some restriction further up.

As for DXIL, we leave it unchanged, as DXIL validation will catch those
cases with a good-enough error message.

The current accepted cases this now blocks were mostly broken later in the
codegen, so this should be an acceptable 'regression'.

Related to #6384, #6220
Fixed #6383, #6382

Signed-off-by: Nathan Gauër <brioche@google.com>