[SPIR-V] Fix struct size computation with bitfields (#6471)
The struct size & alignment computations were wrong when
bitfields were in use. This was due to a mismatch between size
computations at the AST level, and at the SPIR-V level.

This commit adds logic to correctly compute bitfields offsets on
an AST struct.

**note: there are 2 commits in this PR, one that moves code around with
no change, and one that adds the logic. To help you review.**

Fixes #5648

---------

Signed-off-by: Nathan Gauër <brioche@google.com>