Support SetMeshOutputCounts in mesh nodes (#6476)
Allow the usage of SetMeshOutputCounts() in mesh launch node shaders and
no other launch type nodes. This requires identifying the op initially
as permitted in node shaders, but restraining that permission later on.
Given the availability of launch type and opcode information, the
locations for doing that were limited. Future implementation might opt
to encode the launch type in the stage mask or else at least temporarily
use a different opcode for a SetMeshOutputCounts call within a node
shader.

Fixes #6473