Fix recursive zero replacement bottoming out in scalars.
During memcpy replacement, the scalarrepl-param-hlsl pass will replace
certain uses of a zero-initialized global variable.  The pass uses
a recursive algorithm to replace uses of that global by a zero value.
The current code expects the recursion to bottom out in an aggregate type.
But it can bottom out in a scalar integer or scalar float.

This patch fixes the scalar cases.

Lit-based pass tests are included.