Write usage masks for mesh nodes in libraries into RDAT (#6522)
Previously, the step that is typically taken to set the usage mask for
signature elements was not executed, due to mesh nodes being defined in
library shaders. The mask was only written for non-lib shaders. Now that
mesh nodes exist in library shaders, we need to allow usage masks to be
set for mesh nodes, even though they are in library shaders.

This PR adds an alternative case for mesh nodes. In this case, usage
masks are written for all mesh nodes defined in a library shader, by
iterating over all functions declared in a library module. An
accompanying test was written that shows the RDAT now contains the usage
mask, and it is non-null.

Fixes https://github.com/microsoft/DirectXShaderCompiler/issues/6521

---------

Co-authored-by: Tex Riddell <texr@microsoft.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>