[SPIR-V] Remove bitcast from float->int conversion
OpConvertFToU and FToS convert numerically from floating point to
integer, with rounding toward 0.0, matching the defined behavior of
HLSL without any need for an initial bitwidth conversion from floating
point to floating point, which can result in incorrect rounding
behavior (see #6501).

Note that behavior is undefined if the target type is not wide enough to
hold the converted value, however this was also true with an initial
bitcast (because an N-bit FP can hold values outside the range of an
N-bit int), and I can't come up with any case where an intial truncation
would result in correct defined behavior where this straight conversion
would be undefined since the precision will inevitably be lost anyways.

Fixes #6501