[SPIR-V] Fix debug instruction with cbuffer + FXC (#6531)
When the requested constant buffer layout is DX, some type lowering can
become complex (Nx1 matrices for ex). To simplify the lowering, the
backend "clones" those variables:
 - on one end, we expose the correct layout
- but then, we copy this to a local variable to have a different layout
compatible with the operations we usually do on such types.

This means the type can sometimes be an HybridStructType, which is not
completely lowered, or a normal lowerer SPIR-V type. Both should be
allowed, but some codepaths had to be updated to reflect this.

---------

Signed-off-by: Nathan Gauër <brioche@google.com>