Disable non-legacy cbuffer flags
These flags had some effect on load code generation, but never worked
right and were not tested in the simple compiler let alone with drivers.
This removes any notion of the flags beyond option processing and
produces a warning message that the flag is not supported and will be
ignored. The flags are both not listed in help anymore either. A simple
OptionsTest verifies that the warning message is produced for either
flag and existing tests that used the flag are updated or removed.

Fixes #6306