Fix auto-dxilver logic in FileCheckerTest (#6368)
In FileCheckerTest.cpp, the -T target of a %dxc part is checked for
compatibility. However, if you use -Vd, it skips the version check for
the validator. This isn't quite right since the validator version will
be set from the attached validator, unless you explicitly set
-validator-version or use -select-validator internal. Also, if
-validator-version is explicitly set, we should check the validator
against that version, not just the one based on shader model.

This change fixes the initial check logic to only skip validator check
when using -select-validator internal or -validator-version. It also
checks against an explicitly requested -validator-version when an
external validator will be used.

Fixes #6367