prevent static global init breaking up allocas
We try to keep the allocas at the top of the entry block. This
initialization code placed the init instruction just after the alloca it
was initializing, which left it in the middle of the allocas. This
disrupted future attempts to find the end of the allocas and potentially
put instructions referencing allocas before the relevant alloca.