Fix invalid module bitcode when indexing a swizzled bool vector
When indexing a swizzled bool vector, some HLSL-specific code in
EmitCXXMemberOrOperatorMemberCallExpr kicks in to handle the
HLSLVecType. In this case, we’re dealing with an ExtVectorElt because of
the swizzle, so this function creates a GEP, Load, and Store on the
vector. However, we need to truncate the loaded value from i32 to i1,
otherwise we attempt to write an i32 to an i1, which trips the assert.