[NFC][SPIR-V] Refactor SpirvGroupNonUniformOps
A follow-up change will use the PartitionedExclusiveScanNV
GroupOperation, which requires that an additional operand is added to all
GroupNonUniformArithmetic instructions. This means that some of the
SPIR-V opcodes which are currently categorized as unary will become
either unary or binary depending on the GroupOp. Since the arity
distinctions between the OpGroupNonUniform* instructions were already
somewhat arbitrary, I'm prefacing that change by refactoring them into a
single SpirvGroupNonUniformOp instruction type for better reusability.