[SPIR-V] Fix firstbit{high,low} elem check
The code gen for firstbithigh and -low was incorrectly checking the size
of the full (possibly composite) type rather than the element size. This
is now fixed, and I've also switch the check from checking whether the
element type is == 64-bit to != 32-bit, so that it matches the current
limitations of the GLSL extended instructions.

https://registry.khronos.org/SPIR-V/specs/unified1/GLSL.std.450.html#:~:text=%3Cid%3E%0AValue-,FindSMsb,-Signed%2Dinteger%20most

Related to #4702