[SPIR-V] Fix InterlockedMin/Max codegen
RWByteAddressBuffer has overloads for InterlockedMin and InterlockedMax for
signed ints that were failing to compile due to mismatched types in the
generated SPIR-V instruction. This adds the missing cast if necessary.

At the same time, some redundant code is removed from the InterlockedMin/Max
intrinsic non-member functions' codegen to modify the opcode. If it was
necessary in the past, the frontend has since been fixed and it is no longer
necessary. Tests to verify these combinations and the necessary implicit casts
have also been added.

Fixes #3196
Related to #4189, #6254, #5707