Add error when pass not found by Extract IR script
When ExtractIRForPassTest fails to find a pass due to a typo in the name
or other reason, it would manifest as trying to join a None value. This
wasn't too helpful to diagnose the problem. Now it tests for None and
produces an error and a suggestion to identify the problem.