[SPIR-V] Add support for D3D12 descriptor heaps
This commit adds support for Sampler/Resource descriptor heaps in DXC.
Support for those heaps on the SPIR-V side requires no other extension
than SPV_EXT_descriptor_indexing.
On the Vulkan side, the VK_EXT_mutable_descriptor_type will be required
as multiple descriptor types must be allowed on the same binding.

The way this is implemented is by reserving 2 descriptor sets:
 - one for resources/samplers
 - one for counters.

This means the sampler & descriptor heaps are both sharing the same
descriptor set.
Counters are stored in their own set, so their index can match the
associated resource index:

```hlsl
        RWStructuredBuffer a = ResourceDescriptorHeap[2];
        // descriptorSet 0, binding 2 stores the buffer.
        // descriptorSet 1, binding 2 stores the counter.
```

The choice to share descriptor across resource/sampler it because I
felt this was simpler to use on the API side.

The choice to allocate a second descriptor and share the bindings
indices between the resource and the counter was done because it felt
the most intuitive method to use them.

```hlsl
          RWStructuredBuffer a = ResourceDescriptorHeap[0];
          RWStructuredBuffer b = ResourceDescriptorHeap[1];
```

```hlsl
        RWStructuredBuffer a = ResourceDescriptorHeap[0];
        a.IncrementCounter();
        // Not A needs 2 bindings, so the second resource needs to be
        shifted.
        RWStructuredBuffer b = ResourceDescriptorHeap[2];
```

For now, descriptor set indices are fixed: 0 for heaps, 1 for counters.
I think this is fine for now, but we might want to add a flag to change
those values, in case some engine wants to specific descriptors.

Also, I decided to forbid usage of non-explicit bindings when heaps are
in use:

```hlsl
Texture2D textureA; // Forbidden. If the binding set=0, b=0 or set=2, b=0?
float4 main() : SV_Target {
          Texture2D textureB = ResourceDescriptorHeap[0];
```

```hlsl
[[vk::binding(0, 0)]]
Texture2D textureA; // Allowed
float4 main() : SV_Target {
          Texture2D textureB = ResourceDescriptorHeap[0];
```

The reason is usability: I'd assume if you use explicit bindings, you
know what you are doing. Meaning you might want to alias a resource in
the heap using a specific bindings, or declare resources in other
descriptor sets.
But when using the default behavior, the sudden change in behavior might
not be wanted: shall the declared resource alias resources from the
heap? Or shall a 3rd descriptor set be used for those?

This could also be solved through a flag to define the default set, and
allow mixed usage of heaps and global resources. But until there is a
clear ask for this, I wanted to err on the safe side.

Fixes #4255

Signed-off-by: Nathan Gauër <brioche@google.com>