HLMatrixLower: allow exceptions to propagate out
If an exception is thrown, don't block it in the TempOverloadPool
destructor. Allow it to propagate out as a user-visible error.

Explicitly clear the TempOverloadPool before returning from the
HLMatrixLowerPass::runOnModule. In the normal case, when
no exception is thrown, that will still verify that all the
overloads actually have been lowered, and will assert out
if they aren't.