Allow cap and extention attribute on member variables and type aliases
This implements the changes proposed in
https://github.com/microsoft/hlsl-specs/pull/270.

Note that this changes previous behaviour. The inline spir-v spec did
nto allow these attributes on variable declarations, but it was already
implemented. It was implmented in such a way that the capabilities and
extensions were added when the compiler sees a use of the variable. We
now add them when we see the declaration.

This is done because I did not want to try to process the attributes
every time we see the variable used.