[SPIR-V] Fix invalid isnan codegen
IsNan returns a boolean, even is the input-type is a float.
This was working in most cases except:
 - if the layout was not Void
 - if the input type was not a matrix

The first bug is because a bool memory layout/representation is not
specified, and shall never be exposed to externaly-accessible memory.
Hence, if we saw a layout rule != Void, we converted it to a UINT.
When calling isnan, the layout rule should not be propagated as we loose
any layout info.

The second is because our codegen assumed matrix operations returned a
matrix with the same type as the input parameters. In the case of isnan,
this was just wrong.

Fixes #6712

Signed-off-by: Nathan Gauër <brioche@google.com>