Use a real instruction for the sentinel in the Instruction intrusive list
Instructions in a BasicBlock are maintained in a doubly-linked list.
The links are "instrusive" in the sense that with inheritance tricks
the Next and Prev nodes are embedded in the Instruction object itself.

The linked list uses a sentinel object to mark the tail of the list.

Previously, the sentinel was an ilist_half_node<Instruction> which
just consists of a 'Instruction* Prev', but then it was cast to
Instruction*. This is flagged by the undefined behaviour sanitizer
in some cases. Also, it's weird and wrong.

Upstream LLVM has entirely reimplimented the intrusive list to avoid
such problems.  But that's a massive change.

This change uses a real Instruction object as the sentinel.
The instruction is only used for its Next and Prev properties.
I used an UnreachableInst becuase it's small and simple.

Issue: #6446