Remove implicit `this` capture in lambdas
When declaring a lambda with a value-capture default [=, ...], the
this pointer is implicitly captured by value as well. This results
in potentially-unintuitive behavior and has been deprecated in
C++20. It produces a warning in newer versions of clang
(https://reviews.llvm.org/D142639).

This simply makes the implicit captures explicit, preventing the warning. It
does not change the compiled code at all, since it's just removing some
syntactic sugar.