Rework lambdas to explicitly capture variables
Unfortunately, visual studio considers explicit `this` captures
alongside a value-capture default to be an error; this is
currently on the roadmap to be fixed, but with no specific
timeline.
(https://developercommunity.visualstudio.com/t/C20-deprecated-lambda-capture/420398?q=return+value&ftype=idea)

To work around this, we modify all the affected lambdas to
capture all their relevant values directly (including `this`),
rather than using a capture-default.