Enable Living Release Notes (#6772)
Moving the source file for the README.md for github and nuget releases
into the DXC repo. It should be updated with relevant notes for the
upcoming release whenever they are made in main such that they will
already be available when the release is built.

Part of #6697

(cherry picked from commit 8fce06b03c38bc0558dcc3305aa20fe5580da293)