Fix CB vector to scalar array translation generating invalid IR
In the special code to handle the memcpy pattern where a constant buffer
contains a vector array that initializes a local (or static global) scalar
array for use by the shader, an invalid assumption was made that if the
memcpy dest was global, that the src is global as well.

This was not the case, and when expecting to generate constant expressions
to index the src, these generated orphaned instructions instead, leading to
invalid IR.

This fixes the issue by leveraging ReplaceConstantWithInst, and setting the
insertion point for the Builder.  Now, replacement *could* fail, if src
instructions don't dominate replacement uses, so bool for replaced all is
returned from replaceScalarArrayWithVectorArray.

Another issue was that it would replace the dest for the original memcpy
with src along the way.  Now, if we don't replace all uses, this turns the
memcpy into a no-op and any remaining uses are no longer coming from src,
but an undef dest instead.  This was also fixed to skip this replacement,
then clean up this use if all other uses have been successfully replaced.