Release Note for OS-normalized include paths (#6789)
The behavior was changed with #6317 so that regardless of spelling in
the shader, the include path will conform to the host OS style for the
purposes of the include handler. This just adds a release note for that
new behavior.

Fixes #6669

(cherry picked from commit 9fa4618de91a0d80168a13cf71e1fc0355001b7c)