[SPIR-V] Fix flattened SPIR-V variable codegen. (#6756)
When the [branch] annotation is used, switches are converted into an
if/else tree.
Issue arise when declaring a variable into a switch case:
- when flattened, the same variable could be traversed twice in case of
a case fall-through.

In addition, there was a small bug in the flattening logic: only breaks
were stopping the handling of further statements, while early-returns
were ignored. This was not an important bug as it only added more
dead-code, but it was wrong.

Fixes #6718

Signed-off-by: Nathan Gauër <brioche@google.com>