Add option to enable asserts in non-debug and to make asserts trap
If LLVM_ENABLE_ASSERT_IN_NDEBUG is defined, asserts are enabled in
NDEBUG builds. When asserts are enabled, macro ASSERTS_ENABLED is
defined. This is used instead of NDEBUG to wrap assert-only code. Note
that this change does not replace _all_ uses of NDEBUG with
ASSERTS_ENABLED, only the ones required to get the build compiling. The
rest will be handled in future changes.

If LLVM_ASSERT_TRAPS is defined, assert will always trap. Currently,
assert calls RaiseException on Windows, and invokes the standard assert
macro on non-Windows. This macro makes the behaviour well-defined and
consistent across platforms.