Fix warning when use implicit launch type.
By default, the launch type should be set to ‘Broadcast’ when diagnosing barriers.
However, the current behavior sets the default launch type to ‘Invalid,’ resulting in warnings when the launch type is not explicitly specified as an attribute.

To address this issue, we’ll adjust the default setting to ‘Broadcast’ and thereby resolve the problem.