Enable vk-allow-rwstructuredbuffer-arrays, and remove option
We have done all of the development on this feature that we will do.

We cannot handle Multidimentional arrays of these buffers because Vulkan
does not allow it.

We cannot handle arrays that are in a struct if the counter variable is
used. We have tests cases to identify those cases.

Closes https://github.com/microsoft/DirectXShaderCompiler/issues/5440