[SPIRV] Simplify looking for input storage class
The function `DeclResultIdMapper::isInputStorageClass` tries to
determine the storage class for the stage variable by looking at its
SigPoint. This does not always work.

This change simply looks at the storage class that is already associated
with the variable, which should have been set when the StageVar was
created.

Fixes #6959