[SPIR-V] Fix image write to unknown format
When 'unknown' was used in vk::image_format, the value was ignored,
and a format was still guessed.

The fix revealed a bug with Write/ReadWithoutFormat capability gen:
- image formats attributes are applied as a fix-up after initial
  codegen.
- the variable is changed, but not all related operations are fixed.
- capabilities visitor assumed types were coherent, and only looked at
  the OpImageWrite/Read operand type. Not the source variable.
- legalization fixes those incoherences, but runs after the visitor.

Using the capability trimming pass for this is an easy way to solve this
issue.

One test was checking for ReadWithoutFormat, but doesn't require it
since no OpImageRead is done on a Unknown format.

Signed-off-by: Nathan Gauër <brioche@google.com>