[SPIR-V] Fix bad OpLoad on loaded texture
This code path always added an OpLoad since it expected either
an OpAccessChain or a variable.
This is wrong if we get the texture from a called function.

There is one thing I do find weird is the OpAccessChain result
is considered to be an L-value. But I'd expect this to be an r-value.

Fixes #4136

Signed-off-by: Nathan Gauër <brioche@google.com>