[SPIR-V] Fix cast elision with initializer lists
When an a chain of casts were performed in an initializer list,
only the source and destination types were considered.
This means float -> uint was the same as float -> int -> uint.
This however is not correct: intermediate casts can change the result.

Removing the shortcut we took solves this issue.

Fixes #6975

Signed-off-by: Nathan Gauër <brioche@google.com>